backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
rdkit-pypi==2023.9.1
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
celery==5.3.4
redis==5.0.1
sse-starlette==1.8.2
google-generativeai==0.3.2
huggingface-hub==0.20.0
requests==2.31.0

